Move ExportView to its own file
